+++ newswires/client/src/configuration.ts
