+++ newswires/client/src/windowConfigType.ts
